get all serializers in one file
