--- airport/serializers/airplane_serializers.py
@@ -1,14 +1,0 @@
-from rest_framework import serializers
-from airport.models import AirplaneType, Airplane
-
-
-class AirplaneSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Airplane
-        fields = ["id", "name", "rows", "seats_in_row", "airplane_type"]
-
-
-class AirplaneTypeSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = AirplaneType
-        fields = ["id", "name"]
