--- airport/serializers/airport_serializer.py
@@ -1,8 +1,0 @@
-from rest_framework import serializers
-from airport.models import Airport
-
-
-class AirportSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Airport
-        fields = ["id", "name", "closest_big_city"]
