--- airport/serializers/flight_serializers.py
@@ -1,8 +1,0 @@
-from rest_framework import serializers
-from airport.models import Flight
-
-
-class FlightSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Flight
-        fields = ["id", "departure_time", "arrival_time", "route", "airplane", "crew"]
