--- airport/serializers/order_serializers.py
@@ -1,8 +1,0 @@
-from rest_framework import serializers
-from airport.models import Order
-
-
-class OrderSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Order
-        fields = ["id", "created_at", "user"]
