--- airport/serializers/route_serializers.py
@@ -1,8 +1,0 @@
-from rest_framework import serializers
-from airport.models import Route
-
-
-class RouteSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Route
-        fields = ["id", "distance", "source", "destination"]
