--- airport/serializers/ticket_serializers.py
@@ -1,8 +1,0 @@
-from rest_framework import serializers
-from airport.models import Ticket
-
-
-class TicketSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Ticket
-        fields = ["id", "row", "seat", "flight", "order"]
